requirements.txt
anyio==4.9.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==3.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==5.4.0
nest-asyncio==1.6.0
python-dotenv==1.1.1
python-telegram-bot==22.1
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.5.0
